requirements.txt
boto==2.36.0
dj-database-url==0.3.0
Django==1.8
django-autofixture==0.10.1
django-braces==1.4.0
django-crispy-forms==1.4.0
django-crispy-forms-ng==2.0.0
django-extensions==1.5.0
django-filter==0.9.2
django-guardian==1.2.5
django-simple-multitenant==0.1.0
django-storages-redux==1.2.3
django-tenant-schemas==1.5.8
djangorestframework==3.3.2
gunicorn==19.0.0
jsonfield==1.0.3
Markdown==2.6.1
Pillow==2.7.0
psycopg2==2.6.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
Werkzeug==0.10.1
whitenoise==1.0.6
